requirements.txt
httpcore==0.15.0
httpx==0.23.0
idna==3.4
python-telegram-bot==20.0a4
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
tzdata==2022.5
tzlocal==4.2